restructure folder with atomic design
